% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{count_edge_crossings}
\alias{count_edge_crossings}
\title{Count Edge Crossings in Parallel Coordinates}
\usage{
count_edge_crossings(l, r)
}
\arguments{
\item{l}{A numeric vector representing values on the left axis. Must have
the same length as \code{r}.}

\item{r}{A numeric vector representing values on the right axis. Must have
the same length as \code{l}.}
}
\value{
An integer value indicating the total number of edge crossings
between the two axes.
}
\description{
This function calculates the total number of edge crossings in a 2-column
parallel coordinates plot. Each axis represents one of the columns, and
edges (lines) connect corresponding points between the two axes. An edge
crossing occurs when two edges intersect between the axes.
}
\details{
The function checks all pairs of edges formed by the points in \code{l}
and \code{r} to determine whether they cross. Two edges connecting
points (l\link{i}, r\link{i}) and (l\link{j}, r\link{j}) are said to cross if:
\deqn{(l[i] - l[j]) * (r[i] - r[j]) < 0}

\if{html}{\out{<div class="sourceCode">}}\preformatted{     The computational complexity of this function is O(n^2) due to the
     pairwise comparison of all edges. It is suitable for datasets with
     a moderate number of points.
}\if{html}{\out{</div>}}
}
\examples{
# Example with no crossings
left <- c(1, 2, 3)
right <- c(1, 2, 3)
count_edge_crossings(left, right) # Returns 0

# Example with one crossing
left <- c(1, 3)
right <- c(3, 1)
count_edge_crossings(left, right) # Returns 1

# Example with multiple crossings
left <- c(1, 2, 3)
right <- c(3, 1, 2)
count_edge_crossings(left, right) # Returns 2
}
