% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel-package.R
\docType{package}
\name{ggparallel-package}
\alias{ggparallel-package}
\title{ggparallel: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/selkamand/ggparallel}
  \item \url{https://selkamand.github.io/ggparallel/}
  \item Report bugs at \url{https://github.com/selkamand/ggparallel/issues}
}

}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

}
\keyword{internal}
