% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel.R
\name{ggparallel}
\alias{ggparallel}
\title{Parallel Coordinate Plots}
\usage{
ggparallel(
  data,
  col_id = NULL,
  col_colour = NULL,
  order_columns_by = c("appearance", "random", ""),
  verbose = TRUE,
  palette_colour = palette.colors(palette = "Set2"),
  convert_binary_numeric_to_factor = TRUE,
  scaling = c("uniminmax", "none"),
  return = c("plot", "data"),
  options = ggparallel_options()
)
}
\arguments{
\item{return}{}
}
\description{
Parallel Coordinate Plots
}
