% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{ggparallel_options}
\alias{ggparallel_options}
\title{Visual Parameters for ggparallel Plots}
\usage{
ggparallel_options(
  show_legend = TRUE,
  show_legend_titles = FALSE,
  legend_position = c("right", "left", "bottom", "top"),
  legend_title_position = c("top", "bottom", "left", "right"),
  legend_nrow = 4,
  legend_ncol = NULL,
  legend_key_size = 0.3,
  beautify_text = TRUE,
  max_digits_bounds = 1,
  x_axis_text_angle = 45,
  x_axis_text_hjust = 0,
  show_x_axis_text = TRUE,
  show_points = FALSE,
  show_bounds_labels = TRUE,
  show_bounds_rect = TRUE,
  line_alpha = 0.5,
  line_width = NULL,
  line_type = 1,
  interactive_svg_width = NULL,
  interactive_svg_height = NULL
)
}
\arguments{
\item{show_legend}{Display the legend on the plot (flag).}

\item{show_legend_titles}{Display titles for legends (flag).}

\item{legend_position}{Position of the legend ("right", "left", "bottom", "top").}

\item{legend_title_position}{Position of the legend title ("top", "bottom", "left", "right").}

\item{legend_nrow}{Number of rows in the legend (number).}

\item{legend_ncol}{Number of columns in the legend. If set, \code{legend_nrow} should be \code{NULL} (number).}

\item{legend_key_size}{Size of the legend key symbols (number).}

\item{beautify_text}{Beautify y-axis text and legend titles by capitalizing words and adding spaces (flag).}

\item{max_digits_bounds}{Number of digits to round the axis bounds label text to (number)}

\item{line_alpha}{Alpha of line geom (number)}

\item{line_width}{Width of the line geom (number)}

\item{line_type}{Type of line geom (number)}

\item{interactive_svg_width, interactive_svg_height}{Width and height of the interactive graphic region (in inches). Only used when \code{interactive = TRUE}.}
}
\value{
A list of visualization parameters for \code{ggparallel}.
}
\description{
Configures aesthetic and layout settings for plots generated by \code{ggparallel}.
}
