% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel_specific_utils.R
\name{is_crossing}
\alias{is_crossing}
\title{Determine whether a pair of points are crossing in parallel coords}
\usage{
is_crossing(l1, r1, l2, r2)
}
\arguments{
\item{l1}{position of line1 on left axis (number)}

\item{r1}{position of line1 on right axis (number)}

\item{l2}{position of line2 on left axis (number)}

\item{r2}{position of line2 on right axis (number)}
}
\value{
boolean. TRUE if coordinates cross, FALSE if they do not
}
\description{
Determine whether a pair of points are crossing in parallel coords
}
\examples{
is_crossing()
}
