% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{package-ggparallel}
\alias{package-ggparallel}
\alias{package-ggparallel-package}
\title{A package for creating parallel coordinates for categorical data}
\description{
The main function \code{\link{ggparallel}} implements three types of
parallel coordinate plots for categorical data: hammock plots, parallel
sets plots, and common angle plots.
}

