% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{ggparallel-package}
\alias{ggparallel-package}
\title{ggparallel: Variations of Parallel Coordinate Plots for Categorical Data}
\description{
Create hammock plots, parallel sets, and common angle plots with 'ggplot2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/heike/ggparallel/}
  \item \url{https://heike.github.io/ggparallel/}
  \item Report bugs at \url{https://github.com/heike/ggparallel/issues/}
}

}
\author{
\strong{Maintainer}: Heike Hofmann \email{hofmann@mail.iastate.edu} (\href{https://orcid.org/0000-0001-6216-5183}{ORCID})

Authors:
\itemize{
  \item Marie Vendettuoli
}

}
\keyword{internal}
