% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\name{package-ggparallel}
\alias{package-ggparallel}
\title{A package for creating parallel coordinates for categorical data}
\description{
The main function \code{\link{ggparallel}} implements three types of
parallel coordinate plots for categorical data: hammock plots, parallel
sets plots, and common angle plots.
}
